Fix callback order
Hard to catch bug with callbacks (there are linter rules for this but
they produce a lot of noise). If the dependency of a callback hasn't
been defined yet (e.g. `const valueForIdx = ` is still several lines
down), then it will always be undefined at that point and the callback
will never re-calculate.